Struct covid19_vaccine_model_priesemann::vaccination_model::model::AgeGroup

pub struct AgeGroup {
    pub name: String,
    pub M: f64,
    pub influx: f64,
    pub rho: f64,
    pub gamma_I: [f64; 3],
    pub gamma_ICU: [f64; 3],
    pub alpha: [f64; 3],
    pub delta_I: [f64; 3],
    pub delta_ICU: [f64; 3],
    pub min_uptake: f64,
    pub max_uptake: f64,
    pub phase: i32,
}

A collection of all the parameters for a given age group

Fields
name: String

Name of the age group (will be used as the name of the output file when saving the results)

M: f64

Age group population size ($M_i$ in the manuscript)

influx: f64

Daily influx ($\phi_i$ in the manuscript)

rho: f64

E-to-I rate $\rho$ (~1/latent period)

gamma_I: [f64; 3]

Recovery rates in the I compartment (array indices: vaccination status, $\gamma_i^\nu$ in the manuscript)

gamma_ICU: [f64; 3]

Recovery rates in the ICU compartment (array indices: vaccination status, $\gamma_i^{ICU}$ in the manuscript)

alpha: [f64; 3]

I-to-ICU rate (array indices: vaccination status, $\alpha_i^\nu$)

delta_I: [f64; 3]

Fatality rates in the I compartment (array indices: vaccination status, $\delta_i^\nu$)

delta_ICU: [f64; 3]

Fatality rates in the ICU compartment (array indices: vaccination status, $\delta_i^{ICU}$)

min_uptake: f64

Minimal vaccine uptake (as a fraction, used to interpolate)

max_uptake: f64

Maximal vaccine uptake (as a fraction, used to interpolate)

phase: i32

Phase of the vaccination programe that this age group will be prioritised. Used to determine the age stratified vaccinations per week. If phase=-1, they will never get the vaccine (children).

Implementations

impl AgeGroup

pub fn gamma_bar(&self) -> f64

Adds recovery-, ICU- and fatality-rates in the I compartment to $\bar\gamma_i$. Used in the dif. eqs.

pub fn uptake(&self, s: f64) -> f64

Linearily interpolates the uptake between the minimal and maximal uptake. s=0 returns the minimal, s=1 the maximal uptake.