Struct covid19_vaccine_model_priesemann::vaccination_model::model::Model

pub struct Model {
    pub age_groups: Vec<AgeGroup>,
    pub M: f64,
    pub eta0: f64,
    pub kappa0: f64,
    pub tau: f64,
    pub tau_vacc: usize,
    pub vaccinations_per_week_dose1: Vec<Vec<f64>>,
    pub vaccinations_per_week_dose2: Vec<Vec<f64>>,
    pub random_vacc: f64,
    pub N_TTI: f64,
    pub N_test_eff: f64,
    pub N_test_ineff: f64,
    pub N_no_test: f64,
}

A collection of global parameters (especially the vaccination parameters) and all the age groups. Includes the dif. eqs.

Fields
age_groups: Vec<AgeGroup>

A vector of all the age groups

M: f64

Total population size

eta0: f64

~Infection blocking potential of the vaccine ($\eta_0$ in the manuscript)

kappa0: f64

~Vaccine efficacy against severe infections ($\kappa_0$ in the manuscript)

tau: f64

Time between vaccination and developed immunization (time to go from V to S)

tau_vacc: usize

Time between two vaccination doses

vaccinations_per_week_dose1: Vec<Vec<f64>>

In week j, age group i gets vaccinations_per_week_dose1[i][j] first doses

vaccinations_per_week_dose2: Vec<Vec<f64>>

In week j, age group i gets vaccinations_per_week_dose2[i][j] second doses

random_vacc: f64

Fraction of vaccines to distribute randomly across the whole population while prioritising older age groups

N_TTI: f64

Limit of daily infections above which the tracing of test-trace-and-isolate (TTI) breaks down due to overwhelmed authorites

N_test_eff: f64

Limit of daily infections above which effective testing breaks down due to overwhelmed authorites

N_test_ineff: f64

Limit of daily infections above which ineffective testing breaks down due to overwhelmed authorites

N_no_test: f64

Limit of daily infections above which essentially no testing can be performed due to overwhelmed authorites

Implementations

impl Model

pub fn I_eff(&self, state: &Vec<AgeGroupStateVector>) -> f64

Sums up all I compartments of all age groups and vaccinations status weighted by the removal rate from the I compartment, i.e. returns $\sum_{i,\nu}\bar\gamma_i I^\nu_i$. Used for the contagion terms in the dif. eqs.

pub fn slopes(
&self,
t: f64,
h: f64,
R: f64,
state: &Vec<AgeGroupStateVector>,
delayed_R: f64,
delayed_state: &Vec<AgeGroupStateVector>
) -> (Vec<AgeGroupStateVector>, f64)

Implements the dif. eqs. and returns a vector of all slopes d/dt. Needs the current time $t$, the current H-value h (see Solver.H for an explanation) the current $R_t$ value, full system state as well as the delayed $R_{t-\tau}$ and the delayed system state at time $t-\tau$.

Returns: (vector of slopes for all age group compartments, slope for H)

pub fn add_age_group(&mut self, age_group: AgeGroup)

Add an age group to the model. Adds also this age groups population $M_i$ to the total $M$.

fn vaccine_supplies_per_week(week: usize) -> f64

Returns the vaccine supplies $w^T(week)$ for a given week determined by the logistic

$w^T(week)=\frac{11 mio}{1+\exp{-0.17(week-21)}}.$

fn total_vaccination_rates_per_week(week: usize) -> f64

Returns the total doses of vaccination given in a week, i.e. the convoluted supplies, i.e. $\sum_{\tau=0}^2 K[\tau] w^T(week-\tau)$ with $K=[0.6, 0.3, 0.1]$

fn s_for_given_total_uptake(&self, total_uptake: f64) -> f64

Find the linear interpolation parameter s parametrising the individual uptakes for the given total_vaccination_fraction

pub fn prepare_vaccination_rates(&mut self, weeks: usize, total_uptake: f64)

Prepares the weekly vaccination rates per age group (to get the daily rates $f_i^1(t)$ and $f_i^2(t)$ devide by 7). Prepares them for weeks weeks in advance and for a given total uptake.

pub fn vaccinated_between(
&self,
t0: f64,
t1: f64,
age_group: usize
) -> (f64, f64)

Returns how many first and second dose vaccines are distributed between t0 and t1 for a given age group. Used to retrieve the initial conditions.

pub fn N(&self, state: &Vec<AgeGroupStateVector>) -> f64

Calculate the total daily new infections for a given system state (not convoluted by the empirical delay yet)

pub fn ICU_occupancy(&self, state: &Vec<AgeGroupStateVector>) -> f64

Calculate the ICU occupancy for a given system state (Adds all age groups and vaccination status)

pub fn initialize(&mut self)

Prepares the vaccination status dependend transition rates for all the age groups.

The ICU rates remain the same, e.g. $\gamma^{ICU,\nu}_i=\gamma^{ICU,0}_i$

The recovery-, ICU- and fatality-rates in the I compartment get scaled with the vaccine efficacy according to

$\delta_i^{\nu} = (\sqrt{1-\kappa_0})^{\nu}\delta_i $, $\alpha_i^{\nu} = (\sqrt{1-\kappa_0})^{\nu}\alpha_i$ and $ \gamma_i^{\nu}+\delta_i^{\nu}+\alpha_i^{\nu} = \gamma_i$

See the supplementary for more information.

pub fn raw_Rt_to_TTI_corrected(&self, raw_Rt: f64, N: f64) -> f64

Corrects the raw $R_t$ value used in the dif. eqs. by test-trace-and-isolate (TTI) measures, increasing the percieved number of contacts, depending on the current daily infections N.

pub fn raw_Rt_from_TTI_corrected(&self, TTI_Rt: f64, N: f64) -> f64

Calculates the raw $R_t$ value used in the dif. eqs. from the TTI corrected one (see above).